+++ packages/forestadmin-client/src/build-application-services.ts
@@ -4,6 +4,8 @@ import EventsSubscriptionService from './events-subscription';
 import NativeRefreshEventsHandlerService from './events-subscription/native-refresh-events-handler-service';
 import { RefreshEventsHandlerService } from './events-subscription/types';
 import IpWhiteListService from './ip-whitelist';
+import ModelCustomizationFromApiService from './model-customizations/model-customization-from-api';
+import { ModelCustomizationService } from './model-customizations/types';
 import ActionPermissionService from './permissions/action-permission';
 import PermissionService from './permissions/permission-with-cache';
 import RenderingPermissionService from './permissions/rendering-permission';
@@ -31,6 +33,7 @@ export default function buildApplicationServices(
   auth: AuthService;
   eventsSubscription: EventsSubscriptionService;
   eventsHandler: RefreshEventsHandlerService;
+  modelCustomizationService: ModelCustomizationService;
 } {
   const optionsWithDefaults = {
     forestServerUrl: 'https://api.forestadmin.com',
@@ -79,5 +82,9 @@ export default function buildApplicationServices(
     schema: new SchemaService(optionsWithDefaults),
     auth: new AuthService(optionsWithDefaults),
     eventsHandler,
+    modelCustomizationService: new ModelCustomizationFromApiService(
+      forestAdminServerInterface,
+      optionsWithDefaults,
+    ),
   };
 }