+++ packages/forestadmin-client/src/index.ts
@@ -44,6 +44,7 @@ export default function createForestAdminClient(
     auth,
     eventsSubscription,
     eventsHandler,
+    modelCustomizationService,
   } = buildApplicationServices(new ForestHttpApi(), options);
 
   return new ForestAdminClientWithCache(
@@ -57,12 +58,15 @@ export default function createForestAdminClient(
     auth,
     eventsSubscription,
     eventsHandler,
+    modelCustomizationService,
   );
 }
 
 export * from './charts/types';
 export * from './schema/types';
 export { RefreshEventsHandlerService } from './events-subscription/types';
+export * from './model-customizations/types';
+
 export { default as ContextVariablesInjector } from './utils/context-variables-injector';
 export { default as ContextVariables } from './utils/context-variables';
 export { default as ChartHandler } from './charts/chart-handler';