+++ packages/forestadmin-client/src/types.ts
@@ -3,8 +3,9 @@ import type { Chart, QueryChart } from './charts/types';
 import type { Client } from 'openid-client';
 
 import { UserInfo } from './auth/types';
-import { RefreshEventsHandlerService } from './events-subscription/types';
 import { IpWhitelistConfiguration } from './ip-whitelist/types';
+import { ModelCustomization, ModelCustomizationService } from './model-customizations/types';
+import { HttpOptions } from './permissions/forest-http-api';
 import {
   CollectionActionEvent,
   EnvironmentPermissionsV4,
@@ -35,6 +36,7 @@ export interface ForestAdminClient {
   readonly permissionService: PermissionService;
   readonly contextVariablesInstantiator: ContextVariablesInstantiatorInterface;
   readonly chartHandler: ChartHandlerInterface;
+  readonly modelCustomizationService: ModelCustomizationService;
 
   verifySignedActionParameters<TSignedParameters>(signedParameters: string): TSignedParameters;
 
@@ -147,4 +149,5 @@ export interface ForestAdminServerInterface {
   getEnvironmentPermissions: (...args) => Promise<EnvironmentPermissionsV4>;
   getUsers: (...args) => Promise<UserPermissionV4[]>;
   getRenderingPermissions: (renderingId: number, ...args) => Promise<RenderingPermissionV4>;
+  getModelCustomizations: (options: HttpOptions) => Promise<ModelCustomization[]>;
 }