+++ packages/forestadmin-client/test/__factories__/forest-admin-client.ts
@@ -3,8 +3,10 @@ import { Factory } from 'fishery';
 import authServiceFactory from './auth';
 import chartHandlerFactory from './chart/chart-handler';
 import eventsSubscriptionServiceFactory from './events-subscription';
+import eventsHandlerServiceFactory from './events-subscription/native-refresh-events-handler-service';
 import forestAdminClientOptionsFactory from './forest-admin-client-options';
 import ipWhitelistServiceFactory from './ip-whitelist';
+import modelCustomizationServiceFactory from './model-customizations/model-customization-from-api';
 import permissionServiceFactory from './permissions/permission';
 import renderingPermissionsFactory from './permissions/rendering-permission';
 import schemaServiceFactory from './schema';
@@ -33,6 +35,8 @@ const forestAdminClientFactory = ForestAdminClientFactory.define(
       schemaServiceFactory.build(),
       authServiceFactory.build(),
       eventsSubscriptionServiceFactory.build(),
+      eventsHandlerServiceFactory.build(),
+      modelCustomizationServiceFactory.build(),
     ),
 );