+++ packages/_example/CHANGELOG.md
@@ -1,3 +1,43 @@
+## example [1.3.6](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.3.5...example@1.3.6) (2023-05-16)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-sql:** upgraded to 1.6.1
+
+## example [1.3.5](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.3.4...example@1.3.5) (2023-05-10)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-sql:** upgraded to 1.6.0
+
+## example [1.3.5-alpha-webhook-custom-actions.1](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.3.4...example@1.3.5-alpha-webhook-custom-actions.1) (2023-05-10)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.9.0-alpha-webhook-custom-actions.8
+
+## example [1.3.4](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.3.3...example@1.3.4) (2023-05-10)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-sql:** upgraded to 1.5.2
+
 ## example [1.3.3](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.3.2...example@1.3.3) (2023-05-03)
 
 
@@ -9,6 +49,47 @@
 * **@forestadmin/agent:** upgraded to 1.8.13
 * **@forestadmin/datasource-sequelize:** upgraded to 1.3.1
 * **@forestadmin/datasource-sql:** upgraded to 1.5.1
+
+## example [1.3.3-alpha-webhook-custom-actions.4](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.3.3-alpha-webhook-custom-actions.3...example@1.3.3-alpha-webhook-custom-actions.4) (2023-05-03)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.9.0-alpha-webhook-custom-actions.7
+
+## example [1.3.3-alpha-webhook-custom-actions.3](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.3.3-alpha-webhook-custom-actions.2...example@1.3.3-alpha-webhook-custom-actions.3) (2023-05-03)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.9.0-alpha-webhook-custom-actions.6
+
+## example [1.3.3-alpha-webhook-custom-actions.2](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.3.3-alpha-webhook-custom-actions.1...example@1.3.3-alpha-webhook-custom-actions.2) (2023-05-02)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.9.0-alpha-webhook-custom-actions.5
+
+## example [1.3.3-alpha-webhook-custom-actions.1](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.3.2...example@1.3.3-alpha-webhook-custom-actions.1) (2023-04-28)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.9.0-alpha-webhook-custom-actions.4
+* **@forestadmin/datasource-dummy:** upgraded to 1.0.30-alpha-webhook-custom-actions.1
 
 ## example [1.3.2](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.3.1...example@1.3.2) (2023-04-28)
 
@@ -50,6 +131,37 @@
 * **@forestadmin/datasource-sequelize:** upgraded to 1.3.0
 * **@forestadmin/datasource-sql:** upgraded to 1.5.0
 * **@forestadmin/datasource-toolkit:** upgraded to 1.5.0
+
+## example [1.2.21-alpha-webhook-custom-actions.3](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.2.21-alpha-webhook-custom-actions.2...example@1.2.21-alpha-webhook-custom-actions.3) (2023-04-24)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.9.0-alpha-webhook-custom-actions.3
+
+## example [1.2.21-alpha-webhook-custom-actions.2](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.2.21-alpha-webhook-custom-actions.1...example@1.2.21-alpha-webhook-custom-actions.2) (2023-04-20)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.9.0-alpha-webhook-custom-actions.2
+
+## example [1.2.21-alpha-webhook-custom-actions.1](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.2.20...example@1.2.21-alpha-webhook-custom-actions.1) (2023-04-20)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.9.0-alpha-webhook-custom-actions.1
+* **@forestadmin/datasource-dummy:** upgraded to 1.0.28-alpha-webhook-custom-actions.1
 
 ## example [1.2.20](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.2.19...example@1.2.20) (2023-04-19)