+++ packages/datasource-dummy/CHANGELOG.md
@@ -1,3 +1,13 @@
+## @forestadmin/datasource-dummy [1.0.30-alpha-webhook-custom-actions.1](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-dummy@1.0.29...@forestadmin/datasource-dummy@1.0.30-alpha-webhook-custom-actions.1) (2023-04-28)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.8.0-alpha-webhook-custom-actions.1
+
 ## @forestadmin/datasource-dummy [1.0.29](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-dummy@1.0.28...@forestadmin/datasource-dummy@1.0.29) (2023-04-27)
 
 
@@ -23,6 +33,16 @@
 
 * **@forestadmin/datasource-customizer:** upgraded to 1.7.0
 * **@forestadmin/datasource-toolkit:** upgraded to 1.5.0
+
+## @forestadmin/datasource-dummy [1.0.28-alpha-webhook-custom-actions.1](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-dummy@1.0.27...@forestadmin/datasource-dummy@1.0.28-alpha-webhook-custom-actions.1) (2023-04-20)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.7.0-alpha-webhook-custom-actions.1
 
 ## @forestadmin/datasource-dummy [1.0.27](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-dummy@1.0.26...@forestadmin/datasource-dummy@1.0.27) (2023-04-19)