+++ packages/datasource-sql/CHANGELOG.md
@@ -1,3 +1,24 @@
+## @forestadmin/datasource-sql [1.6.1](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-sql@1.6.0...@forestadmin/datasource-sql@1.6.1) (2023-05-16)
+
+
+### Bug Fixes
+
+* **ds-sql:** close the proxy when sequelize instance closes ([#694](https://github.com/ForestAdmin/agent-nodejs/issues/694)) ([d31276c](https://github.com/ForestAdmin/agent-nodejs/commit/d31276cba54b1c4d42e884bdaddb38200cd1364b))
+
+# @forestadmin/datasource-sql [1.6.0](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-sql@1.5.2...@forestadmin/datasource-sql@1.6.0) (2023-05-10)
+
+
+### Features
+
+* **datasource-sql:** enable to proxify the tcp connection by a socks5 proxy ([#690](https://github.com/ForestAdmin/agent-nodejs/issues/690)) ([971a712](https://github.com/ForestAdmin/agent-nodejs/commit/971a71286338ee085d00794dd4fa9f9cf1d05520))
+
+## @forestadmin/datasource-sql [1.5.2](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-sql@1.5.1...@forestadmin/datasource-sql@1.5.2) (2023-05-10)
+
+
+### Bug Fixes
+
+* **datasource-sql:** export missing type ([#691](https://github.com/ForestAdmin/agent-nodejs/issues/691)) ([885de9e](https://github.com/ForestAdmin/agent-nodejs/commit/885de9ecad70c8d9e03afc590c7ecef5aeeadabb))
+
 ## @forestadmin/datasource-sql [1.5.1](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-sql@1.5.0...@forestadmin/datasource-sql@1.5.1) (2023-05-03)